Styles + bugfixes + improvements
Revamped the styles for client & host.
Added some extra features like press enter to send message and a list of users with icons instead of a text peerlist. Changed the way usernames work so they are independt of peerids.
That change makes it so usernames can be the same for now atleast.